+++ src/frontend/src/App.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import "./App.css";
-import {Layout,  theme} from "antd";
+import { Layout, theme } from "antd";
 
-import StudentList from './components/StudentList';
+import StudentList from "./features/student/components/StudentList";
 
 const { Header, Content, Footer } = Layout;
 
@@ -13,13 +13,21 @@ function App() {
   return (
     <div className="App">
       <Layout style={{ minHeight: "100vh" }}>
-          <Header style={{ padding: 0, background: colorBgContainer, textAlign:"center" }}>Student Management App</Header>
-          <Content style={{ margin: "0 16px" }}>
-            <StudentList />
-          </Content>
-          <Footer style={{ textAlign: "center" }}>
-            Powered by Ant Design ©2018. Created by Murtuza Syed
-          </Footer>
+        <Header
+          style={{
+            padding: 0,
+            background: colorBgContainer,
+            textAlign: "center",
+          }}
+        >
+          Student Management App
+        </Header>
+        <Content style={{ margin: "0 16px" }}>
+          <StudentList />
+        </Content>
+        <Footer style={{ textAlign: "center" }}>
+          Powered by Ant Design ©2018. Created by Murtuza Syed
+        </Footer>
       </Layout>
     </div>
   );
