--- src/frontend/src/components/EditStudent.tsx
@@ -1,15 +1,0 @@
-import React, { useState } from "react";
-import { Button, Popconfirm } from "antd";
-
-type EditStudentProps = {
-    onEdit: () => void;
-};
-const EditStudent = ({ onEdit }: EditStudentProps) => {
-    const [open, setOpen] = useState(false);
-    return (
-        <div className={"edit-button"}>
-                <Button onClick={onEdit}>Edit</Button>
-        </div>
-    );
-};
-export default EditStudent;
