+++ src/frontend/src/features/student/components/DeleteStudent.tsx
